confirm delete messages

+++ PianoLessons/ViewModels/ManageCoursesPageViewModel.cs
@@ -91,8 +91,12 @@ public partial class ManageCoursesPageViewModel : ObservableObject
 	[RelayCommand]
 	public async Task RemoveCourse(int courseId)
 	{
-		await service.DeleteCourse(courseId);
-		LoadedCommand.Execute(this);
+		bool confirmDelete = await Application.Current.MainPage.DisplayAlert("Are you sure?", "Do you want to delete this recording?", "Yes", "No");
+		if (confirmDelete)
+		{
+			await service.DeleteCourse(courseId);
+			LoadedCommand.Execute(this);
+		}
 	}
 
 	[RelayCommand]
